Update fwup config to fill out data partition for qemu
This expands the image on disk so that you can have a data partition
with qemu. You can do this manually with dd, but this is way easier. To
avoid wasting too much disk space if using a filesystem that doesn't
support sparse blocks, this reduces the data partition size to 256 MB.
This seems more than ample for qemu use. If written to a physical disk,
the data partition expands to fill it.